requirements.txt
# PySpark + Delta Lake
pyspark==3.5.0
delta-spark==3.1.0

# XGBoost with Spark support
xgboost[spark]==2.0.3  # ✅ เพิ่ม [spark]

# Data processing
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
pyarrow>=9.0.0,<12.0.0

# Optional
jupyterlab==4.0.9
matplotlib==3.8.2
seaborn==0.13.0